% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_connected_constraints.R
\name{add_connected_constraints}
\alias{add_connected_constraints}
\title{Add connected constraint}
\usage{
add_connected_constraints(x)
}
\arguments{
\item{x}{\code{\link{ConservationProblem-class}} object.}
}
\value{
\code{\link{ConservationProblem-class}} object with the constraint
  added to it.
}
\description{
Add constraints to a conservation problem to ensure that all selected
planning units are spatially connected to each other.
}
\examples{
\dontrun{

# create basic problem
p1 <- problem(sim_pu_raster, sim_features) \%>\%
  add_min_set_objective() \%>\%
  add_relative_targets(0.2) \%>\%
  add_default_solver(time_limit=5)

# create problem with added connected constraints
p2 <- p1 \%>\% add_connected_constraints()

# solve problems
s <- stack(solve(p1), solve(p2))

# plot solutions
plot(s, main=c("basic solution", "connected solution"))

}
}
\seealso{
\code{\link{constraints}} for all the available constraints.
}
