% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{array_parameters}
\alias{array_parameters}
\alias{proportion_parameter_array}
\alias{binary_parameter_array}
\alias{numeric_parameter_array}
\title{Array parameters}
\usage{
proportion_parameter_array(name, value, label)

binary_parameter_array(name, value, label)

numeric_parameter_array(name, value, label,
  lower_limit = rep(.Machine$double.xmin, length(value)),
  upper_limit = rep(.Machine$double.xmax, length(value)))
}
\arguments{
\item{name}{\code{character} name of parameter.}

\item{value}{\code{vector} of values.}

\item{label}{\code{character} \code{vector} of labels for each value.}

\item{lower_limit}{\code{vector} of values denoting the minimum acceptable
value for each element in \code{value}. Defaults to the 
smallest possible number on the system.}

\item{upper_limit}{\code{vector} of values denoting the maximum acceptable
value for each element in \code{value}. Defaults to the 
largest  possible number on the system.}
}
\value{
\code{\link{ArrayParameter-class}} object.
}
\description{
Create parameters that consist of multiple numbers. If an attempt is made 
to create a parameter with conflicting settings then an error will be thrown.
}
\details{
Below is a list of parameter generating functions and a brief 
  description of each.
\describe{

  \item{proportion_parameter_array}{a parameter that consists of multiple 
   \code{numeric} values that are between zero and one.}

  \item{numeric_parameter_array}{a parameter that consists of multiple
    \code{numeric} values.}

  \item{binary_parameter_array}{a parameter that consists of multiple
    \code{integer} values that are either zero or one.}

}
}
