% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_matrix.R
\name{boundary_matrix}
\alias{boundary_matrix}
\alias{boundary_matrix-methods}
\alias{boundary_matrix,Raster-method}
\alias{boundary_matrix,SpatialLines-method}
\alias{boundary_matrix,SpatialPoints-method}
\alias{boundary_matrix,SpatialPolygons-method}
\alias{boundary_matrix}
\alias{boundary_matrix}
\alias{boundary_matrix}
\alias{boundary_matrix}
\title{Boundary matrix}
\arguments{
\item{x}{\code{\link[raster]{Raster-class}},
\code{\link[sp]{SpatialLines-class}}, or
\code{\link[sp]{SpatialPolygons-class}} object. If \code{x} is a 
\code{\link[raster]{Raster-class}} object then it must have only one
layer.}

\item{edges}{\code{integer} number of edges that each planning unit has when
 argument to \code{x} is a \code{\link[raster]{Raster-class}} object. 
Defaults to 4. Valid values are 4 and 6.}
}
\value{
\code{\link{Matrix}{dsCMatrix-class}} object.
}
\description{
Generate a boundary matrix describing the shared and exposed edges of 
planning units.
}
\details{
This function returns a \code{\link[Matrix]{dsCMatrix-class}} 
  symmetric sparse matrix. Cells on the off-diagonal indicate the length of 
  the shared boundary between two different planning units. Cells on the 
  diagonal indicate length of a given planning unit's edges that have no 
  neighbors (eg. for edges of planning units found along the 
  coastline). \strong{This function assumes the data are in a coordinate
  system where Euclidean distances accurately describe the proximity
  between two points on the earth}. Thus spatial data in a longitude/latitude
  coordinate system (aka \href{http://spatialreference.org/ref/epsg/wgs-84/}{WGS84})
  should be reprojected to another coordinate system before using this 
  function.
}
