% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connected_matrix.R
\name{connected_matrix}
\alias{connected_matrix}
\alias{connected_matrix,Raster-method}
\alias{connected_matrix,SpatialLines-method}
\alias{connected_matrix,SpatialPoints-method}
\alias{connected_matrix,SpatialPolygons-method}
\alias{connected_matrix}
\alias{connected_matrix}
\alias{connected_matrix}
\alias{connected_matrix}
\title{Connected matrix}
\usage{
connected_matrix(x, directions = 4) # Raster

connected_matrix(x) # SpatialPolygons

connected_matrix(x) # SpatialLines

connected_matrix(x, distance) # SpatialPoints
}
\arguments{
\item{x}{\code{\link[raster]{Raster-class}} or 
\code{\link[sp]{Spatial-class}} object. Note that if \code{x} is a 
\code{\link[raster]{Raster-class}} object then it must have only one
layer.}

\item{directions}{\code{integer} If \code{x} is a
\code{\link[raster]{Raster-class}} object, the number of directions
 in which cells should be connected: 4 (rook's case), 8 (queen's case), 
 16 (knight and one-cell queen moves), or 'bishop' to connect cells with
 one-cell diagonal moves.}

\item{distance}{\code{numeric} If \code{x} is a 
\code{\link{SpatialPoints-class}} object, the distance that planning units 
have to be within in order to qualify as being connected.}
}
\value{
\code{\link[Matrix]{dsCMatrix-class}} object.
}
\description{
Create a matrix showing which planning units are spatially connected.
}
\details{
This function returns a \code{\link[Matrix]{dgCMatrix-class}} sparse
  matrix. Cells along the off-diagonal indicate if two planning units are 
  connected. Cells along the diagonal are zero to reduce memory consumption.
}
