% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{distribute_load}
\alias{distribute_load}
\title{Distribute load}
\usage{
distribute_load(x, n = get_number_of_threads())
}
\arguments{
\item{x}{\code{integer} number of item to process.}

\item{n}{\code{integer} number of threads.}
}
\value{
\code{list} object.
}
\description{
Utility function for distributing computations among a pool of workers
for parallel processing.
}
\details{
This function returns a \code{list} containing an element for
  each worker. Each element contains a \code{integer} \code{vector} 
  specifying the indices that the worker should process.
}
\seealso{
\code{\link{get_number_of_threads}}, 
  \code{\link{set_number_of_threads}}, \code{\link{is.parallel}}.
}
