% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_extract.R
\name{fast_extract}
\alias{fast_extract}
\alias{fast_extract,Raster,SpatialLines-method}
\alias{fast_extract,Raster,SpatialPoints-method}
\alias{fast_extract,Raster,SpatialPolygons-method}
\alias{fast_extract}
\alias{fast_extract}
\alias{fast_extract}
\title{Fast extract}
\usage{
fast_extract(x, y) # Raster, SpatialPolygons

fast_extract(x, y) # Raster, SpatialLines

fast_extract(x, y) # Raster, SpatialPoints
}
\arguments{
\item{x}{\code{\link[raster]{Raster-class}} object.}

\item{y}{\code{\link[sp]{Spatial-class}} object.}

\item{...}{additional arguments passed to \code{\link[raster]{extract}}.}
}
\value{
\code{data.frame}, \code{matrix}, or \code{list} object 
  depending on the arguments.
}
\description{
Extract data from a \code{\link[raster]{Raster-class}} object from a 
\code{\link[sp]{Spatial-class}} object using performance enhancing tricks.
}
\details{
Spatial analyses will be conducted using the 
  \code{\link[velox]{velox}} package if it is installed. Additionally, 
  multiple threads can be used to speed up computation using the 
  \code{\link{set_number_of_threads}} function.
}
\seealso{
\code{\link[raster]{extract}},
  \code{\link[velox]{VeloxRaster_extract}}.
}
