% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{is.parallel}
\alias{is.parallel}
\title{Is parallel?}
\usage{
is.parallel()
}
\value{
\code{logical} indicating if parallel computations will be
  performed in parallel where possible.
}
\description{
This function determines if parallel processing capabilities have been 
initialized.
}
\seealso{
\code{\link{set_number_of_threads}}, 
  \code{\link{get_number_of_threads}}.
}
