% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Parameters}
\usage{
parameters(...)
}
\arguments{
\item{...}{\code{\link{ArrayParameter-class}} and/or 
\code{\link{ScalarParameter-class}}) objects to store in collection.}
}
\value{
\code{\link{Parameters-class}} object.
}
\description{
Create a collection of parameter objects.
}
\seealso{
\code{\link{array_parameters}}, code{\link{scalar_parameters}}.
}
