% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target-proto.R
\name{Target-class}
\alias{Target-class}
\title{Target prototype
 
This prototype is used to represent the targets used when making a
prioritization. This prototype inherits from the 
\code{\link{ConservationModifier-class}}. \strong{This class represents a 
recipe, to actually add targets to a planning problem, see the help page on
\code{\link{targets}}. Only experts should use this class directly.}}
\description{
Target prototype
 
This prototype is used to represent the targets used when making a
prioritization. This prototype inherits from the 
\code{\link{ConservationModifier-class}}. \strong{This class represents a 
recipe, to actually add targets to a planning problem, see the help page on
\code{\link{targets}}. Only experts should use this class directly.}
}
\seealso{
\code{\link{ConservationModifier-class}}.
}
