% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersecting_units.R
\name{intersecting_units}
\alias{intersecting_units}
\alias{intersecting_units,Raster,Raster-method}
\alias{intersecting_units,Raster,Spatial-method}
\alias{intersecting_units,Spatial,Raster-method}
\alias{intersecting_units,Spatial,Spatial-method}
\alias{intersecting_units}
\alias{intersecting_units}
\alias{intersecting_units}
\alias{intersecting_units}
\alias{intersecting_units}
\title{Find intersecting units}
\usage{
intersecting_units(x, y, ...) # x=Raster, y=Raster

intersecting_units(x, y, ...) # x=Spatial, y=Spatial

intersecting_units(x, y, ...) # x=Raster, y=Spatial

intersecting_units(x, y, ...) # x=Spatial, y=Raster

intersecting_units(x, y, ...) # x=data.frame, y=ANY
}
\arguments{
\item{x}{\code{\link{Spatial-class}} or \code{\link{Raster-class}} object.}

\item{y}{\code{\link{Spatial-class}} or \code{\link{Raster-class}} object.}

\item{...}{not used.}
}
\value{
\code{integer} indices of the units in \code{x} that intersect with 
  \code{y}.
}
\description{
Determine the units in a given spatial data object that intersect
with any of the units in another spatial data object.
}
\details{
The \code{\link{set_number_of_threads}} can be used to distribute
  computations among multiple threads and potentially reduce run time.
}
\seealso{
\code{\link{fast_extract}}, \code{\link{set_number_of_threads}},
  \code{\link{get_number_of_threads}}.
}
