% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan_problem.R
\name{marxan_problem}
\alias{marxan_problem}
\alias{marxan_problem.default}
\alias{marxan_problem.data.frame}
\alias{marxan_problem.character}
\title{Marxan conservation problem}
\usage{
marxan_problem(x, ...)

\method{marxan_problem}{default}(x, features, targets,
  targets_type = c("relative", "absolute"), locked_in = waiver(),
  locked_out = waiver(), penalty = 0, edge_factor = 0.5, ...)

\method{marxan_problem}{data.frame}(x, spec, puvspr, bound = NULL, blm = 0,
  asymmetric_connectivity = FALSE, ...)

\method{marxan_problem}{character}(x, ...)
}
\arguments{
\item{x}{\code{character} file path for Marxan input file (typically
called \code{'input.dat'}) or a \code{\link[sp]{Spatial-class}} or 
\code{\link[raster]{Raster-class}} object containing planning unit data.}

\item{...}{not used}

\item{features}{\code{\link[raster]{Raster-class}} object containing
data on the distribution of features across the study area.}

\item{targets}{\code{numeric} amount of each feature to be represented
in the solution.}

\item{targets_type}{\code{character} name indicating if the \code{targets}
are expressed as \code{'relative'} (eg. \code{0.2} meaning that 20 \% of a 
feature needs to be conserved), or \code{'absolute'} (eg. \code{200} 
meaning that 200 units of a feature need to be conserved) amounts.}

\item{locked_in}{Object that determines which planning units that should be 
locked in. See details for more information.}

\item{locked_out}{Object that determines which planning units that should be 
locked out. See details for more information.}

\item{penalty}{\code{numeric} penalty for exposed edges. This
is equivalent to the
\href{http://marxan.net/downloads/uq_marxan_web_2/module2.html}{boundary length modifier (BLM)}
parameter in \href{http://marxan.net}{marxan.net}.}

\item{edge_factor}{\code{numeric} proportion to scale edges that do
not have any neighboring planning units. For example, an edge factor
of \code{0.5} is commonly used for planning units along the coast line.}
}
\value{
\code{\link{ConservationProblem-class}} object.
}
\description{
Create a Marxan conservation problem. Although users are encouraged to 
build and tailor conservation planning problems to suit their own needs, 
sometimes it is easier to use a more familiar "canned" approach. This 
function provides a convenient wrapper for generating
and solving Marxan-style conservation problems. If users already have
their conservation planning data in the Marxan input format, this function
can also be used to read Marxan data files and solve the Marxan-style 
problems using exact algorithm solvers.
}
\examples{
\donttest{

# create Marxan problem using spatial data
data(sim_pu_raster, sim_features)
p1 <- marxan_problem(sim_pu_raster, features=sim_features, targets=0.2, 
                     targets_type='relative', penalty=1, 
                     edge_factor=0.5)

# solve problem
s1 <- solve(p1)

# show solution
plot(s1)

# create marxan problem using marxan input files
input_file <- system.file('extdata/input.dat', package='prioritizrutils')
p2 <- marxan_problem(input_file)

# solve problem
s2 <- solve(p2)

# count number of selected planning units in solution
print(sum(s2))

}

}
